+++ clients/admin-ui/cypress/e2e/config-wizard.cy.ts
@@ -6,7 +6,7 @@ describe("Config Wizard", () => {
   beforeEach(() => {
     cy.login();
     cy.intercept("GET", "/api/v1/organization/*", {
-      fixture: "organization.json",
+      fixture: "organizations/default_organization.json",
     }).as("getOrganization");
   });