+++ clients/admin-ui/src/features/common/nav/v2/hooks.ts
@@ -1,10 +1,13 @@
+import {
+  configureNavGroups,
+  findActiveNav,
+  NAV_CONFIG,
+} from "@fidesui/components";
 import { useMemo } from "react";
 
 import { useAppSelector } from "~/app/hooks";
 import { useFeatures } from "~/features/common/features";
 import { selectThisUsersScopes } from "~/features/user-management";
-
-import { configureNavGroups, findActiveNav, NAV_CONFIG } from "./nav-config";
 
 export const useNav = ({ path }: { path: string }) => {
   const features = useFeatures();
@@ -15,8 +18,7 @@ export const useNav = ({ path }: { path: string }) => {
       configureNavGroups({
         config: NAV_CONFIG,
         hasPlus: features.plus,
-        hasAccessToPrivacyRequestConfigurations:
-          features.flags.privacyRequestsConfiguration,
+        flags: features.flags,
         userScopes,
       }),
     [features, userScopes]