+++ tests/ctl/cli/test_cli.py
@@ -825,7 +825,6 @@ class TestGenerate:
         test_cli_runner: CliRunner,
         tmpdir: LocalPath,
     ) -> None:
-
         tmp_output_file = tmpdir.join("dataset.yml")
         config_data = os.getenv("BIGQUERY_CONFIG", "e30=")
         config_data_decoded = loads(
@@ -875,7 +874,6 @@ class TestGenerate:
         test_cli_runner: CliRunner,
         tmpdir: LocalPath,
     ) -> None:
-
         tmp_output_file = tmpdir.join("dataset.yml")
         tmp_keyfile = tmpdir.join("bigquery.json")
         config_data = os.getenv("BIGQUERY_CONFIG", "e30=")
@@ -911,6 +909,7 @@ def credentials_path(tmp_path_factory) -> str:
     return str(credentials_path)
 
 
+@pytest.mark.integration
 class TestUser:
     """
     Test the "user" command group.
@@ -918,7 +917,6 @@ class TestUser:
     Most tests rely on previous tests.
     """
 
-    @pytest.mark.unit
     def test_user_login_provide_credentials(
         self, test_config_path: str, test_cli_runner: CliRunner, credentials_path: str
     ) -> None:
@@ -941,7 +939,75 @@ class TestUser:
         print(result.output)
         assert result.exit_code == 0
 
-    @pytest.mark.unit
+    def test_user_login_env_var_failed(
+        self, test_config_path: str, test_cli_runner: CliRunner, credentials_path: str
+    ) -> None:
+        """
+        Test logging in as a user with a provided username and password
+        provided via env vars, but the username is invalid.
+        """
+        print(credentials_path)
+        result = test_cli_runner.invoke(
+            cli,
+            [
+                "-f",
+                test_config_path,
+                "user",
+                "login",
+            ],
+            env={
+                "FIDES_CREDENTIALS_PATH": credentials_path,
+                "FIDES__USER__USERNAME": "fakeuser",
+                "FIDES__USER__PASSWORD": "Testpassword1!",
+            },
+        )
+        print(result.output)
+        assert result.exit_code == 1
+
+    def test_user_login_env_var_password(
+        self, test_config_path: str, test_cli_runner: CliRunner, credentials_path: str
+    ) -> None:
+        """
+        Test logging in as a user with a provided username but password
+        provided via env vars.
+        """
+        print(credentials_path)
+        result = test_cli_runner.invoke(
+            cli,
+            ["-f", test_config_path, "user", "login", "-u", "root_user"],
+            env={
+                "FIDES_CREDENTIALS_PATH": credentials_path,
+                "FIDES__USER__PASSWORD": "Testpassword1!",
+            },
+        )
+        print(result.output)
+        assert result.exit_code == 0
+
+    def test_user_login_env_var_credentials(
+        self, test_config_path: str, test_cli_runner: CliRunner, credentials_path: str
+    ) -> None:
+        """
+        Test logging in as a user with a provided username and password
+        provided via env vars.
+        """
+        print(credentials_path)
+        result = test_cli_runner.invoke(
+            cli,
+            [
+                "-f",
+                test_config_path,
+                "user",
+                "login",
+            ],
+            env={
+                "FIDES_CREDENTIALS_PATH": credentials_path,
+                "FIDES__USER__USERNAME": "root_user",
+                "FIDES__USER__PASSWORD": "Testpassword1!",
+            },
+        )
+        print(result.output)
+        assert result.exit_code == 0
+
     def test_user_create(
         self, test_config_path: str, test_cli_runner: CliRunner, credentials_path: str
     ) -> None:
@@ -954,9 +1020,7 @@ class TestUser:
                 test_config_path,
                 "user",
                 "create",
-                "-u",
                 "newuser",
-                "-p",
                 "Newpassword1!",
             ],
             env={"FIDES_CREDENTIALS_PATH": credentials_path},
@@ -986,7 +1050,6 @@ class TestUser:
         assert set(total_scopes) == set(SCOPE_REGISTRY)
         assert roles == [OWNER]
 
-    @pytest.mark.unit
     def test_user_permissions_valid(
         self, test_config_path: str, test_cli_runner: CliRunner, credentials_path: str
     ) -> None:
@@ -1000,7 +1063,6 @@ class TestUser:
         print(result.output)
         assert result.exit_code == 0
 
-    @pytest.mark.unit
     def test_get_self_user_permissions(
         self, test_config_path, test_cli_runner, credentials_path
     ) -> None:
@@ -1053,7 +1115,6 @@ class TestUser:
         )
         assert systems == []
 
-    @pytest.mark.unit
     def test_get_other_user_perms_and_systems(
         self, test_config_path, test_cli_runner, credentials_path, system_manager
     ) -> None:
@@ -1087,7 +1148,6 @@ class TestUser:
         )
         assert systems == [system_manager.systems[0].fides_key]
 
-    @pytest.mark.unit
     def test_user_permissions_not_found(
         self, test_config_path: str, test_cli_runner: CliRunner, credentials_path: str
     ) -> None: