+++ tests/ops/service/messaging/message_dispatch_service_test.py
@@ -100,7 +100,6 @@ class TestMessageDispatchService:
     def test_email_dispatch_mailgun_config_not_found(
         self, mock_mailgun_dispatcher: Mock, db: Session
     ) -> None:
-
         with pytest.raises(MessageDispatchException) as exc:
             dispatch_message(
                 db=db,
@@ -123,7 +122,6 @@ class TestMessageDispatchService:
     def test_email_dispatch_mailgun_config_no_secrets(
         self, mock_mailgun_dispatcher: Mock, db: Session
     ) -> None:
-
         messaging_config = MessagingConfig.create(
             db=db,
             data={
@@ -277,7 +275,6 @@ class TestMessageDispatchService:
     def test_sms_dispatch_twilio_success(
         self, mock_twilio_dispatcher: Mock, db: Session, messaging_config_twilio_sms
     ) -> None:
-
         dispatch_message(
             db=db,
             action_type=MessagingActionType.SUBJECT_IDENTITY_VERIFICATION,
@@ -315,7 +312,6 @@ class TestMessageDispatchService:
     def test_sms_dispatch_twilio_config_not_found(
         self, mock_twilio_dispatcher: Mock, db: Session
     ) -> None:
-
         with pytest.raises(MessageDispatchException) as exc:
             dispatch_message(
                 db=db,
@@ -339,7 +335,6 @@ class TestMessageDispatchService:
     def test_sms_dispatch_twilio_config_no_secrets(
         self, mock_mailgun_dispatcher: Mock, db: Session
     ) -> None:
-
         messaging_config = MessagingConfig.create(
             db=db,
             data={