+++ CHANGELOG.md
@@ -17,7 +17,13 @@ The types of changes are:
 
 ## [Unreleased](https://github.com/ethyca/fides/compare/2.9.1...main)
 
+### Added
+* Allow users to configure their username and password via the config file [#2884](https://github.com/ethyca/fides/pull/2884)
+* Add authentication to the `masking` endpoints as well as accompanying scopes [#2909](https://github.com/ethyca/fides/pull/2909)
+* Add an Organization Management page (beta) [#2908](https://github.com/ethyca/fides/pull/2908)
+
 ### Changed
+
 * Improved standard layout for large width screens and polished misc. pages [#2869](https://github.com/ethyca/fides/pull/2869)
 * Changed UI paths in the admin-ui [#2869](https://github.com/ethyca/fides/pull/2892)
   * `/add-systems/new` --> `/add-systems/manual`
@@ -26,17 +32,23 @@ The types of changes are:
 * Deprecated adding scopes to users directly; you can only add roles. [#2848](https://github.com/ethyca/fides/pull/2848/files)
 * Changed About Fides page to say "Fides Core Version:" over "Version". [#2899](https://github.com/ethyca/fides/pull/2899)
 * Polish Admin UI header & navigation [#2897](https://github.com/ethyca/fides/pull/2897)
+* Give new users a "viewer" role by default [#2900](https://github.com/ethyca/fides/pull/2900)
+* Tie together save states for user permissions and systems [#2913](https://github.com/ethyca/fides/pull/2913)
+* Removing payment types from Stripe connector params [#2915](https://github.com/ethyca/fides/pull/2915)
+
 
 ### Fixed
+
 * Restricted Contributors from being able to create Owners [#2888](https://github.com/ethyca/fides/pull/2888)
-
-### Fixed
+* Remove policy key from Privacy Center submission modal [#2912](https://github.com/ethyca/fides/pull/2912)
+* Allow multiple data uses as long as their processing activity name is different [#2905](https://github.com/ethyca/fides/pull/2905)
 * Allow for dynamic aspect ratio for logo on Privacy Center 404 [#2895](https://github.com/ethyca/fides/pull/2895)
 * Allow for dynamic aspect ratio for logo on consent page [#2895](https://github.com/ethyca/fides/pull/2895)
 
 ## [2.9.1](https://github.com/ethyca/fides/compare/2.9.0...2.9.1)
 
 ### Added
+
 * Added Attentive erasure email connector [#2782](https://github.com/ethyca/fides/pull/2782)
 
 ### Changed